+++ pulsar-io/kafka/src/main/java/org/apache/pulsar/io/kafka/ByteBufferSchemaWrapper.java
@@ -1,0 +1,74 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.pulsar.io.kafka;
+
+import lombok.Value;
+import org.apache.pulsar.client.api.Schema;
+import org.apache.pulsar.common.schema.SchemaInfo;
+
+import java.nio.ByteBuffer;
+import java.util.function.Supplier;
+
+/**
+ * This is a ByteBuffer schema that reports SchemaInfo from another Schema instance.
+ */
+@Value
+public class ByteBufferSchemaWrapper implements Schema<ByteBuffer> {
+    private final Supplier<SchemaInfo> original;
+
+    public ByteBufferSchemaWrapper(Schema original) {
+        this.original = original::getSchemaInfo;
+    }
+
+    public ByteBufferSchemaWrapper(SchemaInfo info) {
+        this.original = () -> info;
+    }
+
+    @Override
+    public byte[] encode(ByteBuffer message) {
+        return getBytes(message);
+    }
+
+    @Override
+    public SchemaInfo getSchemaInfo() {
+        return original.get();
+    }
+
+    @Override
+    public Schema<ByteBuffer> clone() {
+        return this;
+    }
+
+    @Override
+    public ByteBuffer decode(byte[] bytes, byte[] schemaVersion) {
+        throw new UnsupportedOperationException();
+    }
+
+    public String toString() {
+        return "{schema wrapper for "+ original +"}";
+    }
+
+    static byte[] getBytes(ByteBuffer buffer) {
+        buffer.mark();
+        byte[] avroEncodedData = new byte[buffer.remaining()];
+        buffer.get(avroEncodedData);
+        buffer.reset();
+        return avroEncodedData;
+    }
+}
